Move all display classes into the EntityEmbedDisplay namespace.
