+++ entity_embed.services.yml
@@ -1,4 +1,4 @@
 services:
   plugin.manager.entity_embed.display:
-    class: Drupal\entity_embed\EntityEmbedDisplayManager
+    class: Drupal\entity_embed\EntityEmbedDisplay\EntityEmbedDisplayManager
     arguments: ['@container.namespaces', '@cache.discovery', '@language_manager', '@module_handler', '@string_translation']
